requirements.txt
agora==1.2.3
anyio==3.6.2
argh==0.28.1
asgiref==3.5.2
asn1crypto==1.5.1
autocommand==2.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
contourpy==1.0.7
crispy==0.7.4
crispy-bootstrap5==0.7
cryptography==40.0.1
cycler==0.11.0
dateutils==0.6.12
defusedxml==0.7.1
Django==4.1.1
django-agora==0.4
django-allauth==0.54.0
django-auto-logout==0.5.1
django-ckeditor==6.5.1
django-countries==7.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fabio==2023.4.0
fastapi==0.95.0
fonttools==4.39.3
h11==0.14.0
h5py==3.8.0
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
more-itertools==9.0.0
ndg-httpsclient==0.5.1
numpy==1.24.2
oauthlib==3.2.2
onyx.core==1.0.5
oscrypto==1.3.0
packaging==23.0
Pillow==9.2.0
pluggy==1.0.0
psycopg2==2.9.6
pusher==3.3.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.5.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
razorpay==1.3.0
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.1
silx==1.1.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.26.1
tinycss2==1.1.1
typing_extensions==4.4.0
tzdata==2022.5
uritools==4.0.0
urllib3==1.26.12
uvicorn==0.21.1
webencodings==0.5.1
zc.lockfile==2.0
